monitoring/telegram_bot/requirements.txt
requests==2.32.3
docker==7.1.0
psycopg2-binary==2.9.9
